New implementation with mmap and munmap
We ran into the issue of OOM for mmap regions.
After printing stats about the memory usage, I realized that most of the
memory is allocated and then freed and it seemed stupid to not implement
free.
I got a bitmap-based tracker of free pages (from chat gpt that
I adapted).
We now seem to be able to satisfy even 100k requests